[react-native] Refactor "QuilttConnector" for better visibility
