+++ packages/react-native/package.json
@@ -37,12 +37,16 @@
   "dependencies": {
     "@honeybadger-io/core": "6.6.0",
     "@quiltt/core": "workspace:*",
-    "@quiltt/react": "workspace:*"
+    "@quiltt/react": "workspace:*",
+    "base-64": "1.0.0",
+    "expo-linking": "6.3.1",
+    "react-native-url-polyfill": "2.0.0",
+    "react-native-webview": "13.10.2"
   },
   "devDependencies": {
     "@apollo/client": "3.9.9",
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
-    "@types/base-64": "0.1.0",
+    "@types/base-64": "1.0.2",
     "@types/node": "20.12.7",
     "@types/react": "18.2.73",
     "@types/react-native": "0.72.5",
@@ -62,6 +66,7 @@
   },
   "peerDependencies": {
     "base-64": "*",
+    "expo-linking": "^6.0.0",
     "react": "^18.2.0",
     "react-native": "^0.72.5",
     "react-native-url-polyfill": "*",
