--- packages/react-native/src/utils/connector/handleOAuthUrl.ts
@@ -1,9 +1,0 @@
-import { Linking } from 'react-native'
-
-export const handleOAuthUrl = (oauthUrl: URL) => {
-  if (oauthUrl.protocol !== 'https:') {
-    console.log(`handleOAuthUrl - Skipping non https url - ${oauthUrl.href}`)
-    return
-  }
-  Linking.openURL(oauthUrl.href)
-}
