--- packages/react-native/src/utils/connector/index.ts
@@ -1,2 +1,0 @@
-export * from './checkConnectorUrl'
-export * from './handleOAuthUrl'
