+++ packages/react-native/src/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './connector'
 export * from './error'
